some refactoring and renames
