+++ public/index.html
@@ -35,7 +35,7 @@
           <section class="trip-events">
             <h2 class="visually-hidden">Trip events</h2>
 
-            <ul class="trip-events__list"></ul>
+
           </section>
       </div>
     </main>
