+++ src/view/events-item-form.js
@@ -1,4 +1,4 @@
-export const createNewEvensItemForm = () => (
+export const createEventsItemForm = () => (
   `<li class="trip-events__item">
     <form class="event event--edit" action="#" method="post">
       <header class="event__header">
