+++ src/view/events-item.js
