+++ src/view/events-list.js
