+++ src/view/trip-controls.js
@@ -1,4 +1,4 @@
-export const createSiteMenuTemplate = () => (
+export const createTripControls = () => (
   `<nav class="trip-controls__trip-tabs  trip-tabs">
     <a class="trip-tabs__btn  trip-tabs__btn--active" href="#">Table</a>
     <a class="trip-tabs__btn" href="#">Stats</a>
