+++ src/view/trip-info.js
@@ -1,4 +1,4 @@
-export const createTripInfoTemplate = () => (
+export const createTripInfo = () => (
   `<section class="trip-main__trip-info  trip-info">
      <div class="trip-info__main">
         <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
